+++ sylloge/base.py
@@ -1,57 +1,106 @@
-import pathlib
 import logging
+import pathlib
 from abc import abstractmethod
 from dataclasses import dataclass
-from typing import Optional, Sequence, Union
+from typing import (
+    TYPE_CHECKING,
+    Generic,
+    Literal,
+    Optional,
+    Sequence,
+    TypeVar,
+    Union,
+    overload,
+)
 
 import pandas as pd
 import pystow
 from pystow.utils import read_zipfile_csv
+from slugify import slugify
 
+from .dask import read_dask_df_archive_csv
 from .typing import EA_SIDES, LABEL_HEAD, LABEL_RELATION, LABEL_TAIL
-from .utils import fix_dataclass_init_docs
-from .utils import load_from_rdf
+from .utils import fix_dataclass_init_docs, load_from_rdf
 
-BASE_DATASET_MODULE = pystow.module("sylloge")
+BASE_DATASET_KEY = "sylloge"
+
+BASE_DATASET_MODULE = pystow.module(BASE_DATASET_KEY)
+
+T = TypeVar("T")
+
+BACKEND_LITERAL = Literal["pandas", "dask"]
+
+if TYPE_CHECKING:
+    import dask.dataframe as dd
 
 logger = logging.getLogger(__name__)
 
+
 @fix_dataclass_init_docs
 @dataclass
-class TrainTestValSplit:
+class TrainTestValSplit(Generic[T]):
     """Dataclass holding split of gold standard entity links."""
 
     #: entity links for training
-    train: pd.DataFrame
+    train: T
     #: entity links for testing
-    test: pd.DataFrame
+    test: T
     #: entity links for validation
-    val: pd.DataFrame
+    val: T
 
 
 @fix_dataclass_init_docs
 @dataclass
-class EADataset:
+class EADataset(Generic[T]):
     """Dataclass holding information of the alignment class."""
 
     #: relation triples of left knowledge graph
-    rel_triples_left: pd.DataFrame
+    rel_triples_left: T
     #: relation triples of right knowledge graph
-    rel_triples_right: pd.DataFrame
+    rel_triples_right: T
     #: attribute triples of left knowledge graph
-    attr_triples_left: pd.DataFrame
+    attr_triples_left: T
     #: attribute triples of right knowledge graph
-    attr_triples_right: pd.DataFrame
+    attr_triples_right: T
     #: gold standard entity links of alignment
-    ent_links: pd.DataFrame
+    ent_links: T
     #: optional pre-split folds of the gold standard
-    folds: Optional[Sequence[TrainTestValSplit]] = None
+    folds: Optional[Sequence[TrainTestValSplit[T]]] = None
+
+    def _canonical_name(self) -> str:
+        raise NotImplementedError
+
+    @property
+    def canonical_name(self) -> str:
+        """A canonical name for this dataset instance.
+
+        This includes all the necessary information
+        to distinguish this specific dataset as string.
+        This can be used e.g. to create folders with this
+        dataset name to store results.
+
+        :return: concise string representation for this dataset instance
+        """
+        name = self._canonical_name
+        assert isinstance(name, str)  # for mypy
+        return slugify(name, separator="_")
+
+    def _param_repr(self) -> str:
+        raise NotImplementedError
+
+    @property
+    def _statistics(self) -> str:
+        if hasattr(self.rel_triples_left, "__len__"):
+            return f"rel_triples_left={len(self.rel_triples_left)}, rel_triples_right={len(self.rel_triples_right)}, attr_triples_left={len(self.attr_triples_left)}, attr_triples_right={len(self.attr_triples_right)}, ent_links={len(self.ent_links)}, folds={len(self.folds) if self.folds else None}"  # type: ignore
+        else:
+            unknown = "unknown_len"
+            return f"rel_triples_left={unknown}, rel_triples_right={unknown}, attr_triples_left={unknown}, attr_triples_right={unknown}, ent_links={unknown}, folds={unknown if self.folds else None}"
 
     def __repr__(self) -> str:
-        return f"{self.__class__.__name__}(rel_triples_left={len(self.rel_triples_left)}, rel_triples_right={len(self.rel_triples_right)}, attr_triples_left={len(self.attr_triples_left)},attr_triples_right={len(self.attr_triples_right)},ent_links={len(self.ent_links)})"
+        return f"{self.__class__.__name__}({self._param_repr()}{self._statistics})"
 
 
-class ZipEADataset(EADataset):
+class ZipEADataset(EADataset[pd.DataFrame]):
     """Dataset created from zip file which is downloaded."""
 
     def __init__(
@@ -63,6 +112,7 @@ class ZipEADataset(EADataset):
         file_name_attr_triples_left: str = "attr_triples_1",
         file_name_attr_triples_right: str = "attr_triples_2",
         file_name_ent_links: str = "ent_links",
+        backend: BACKEND_LITERAL = "pandas",
     ):
         """Initialize ZipEADataset.
 
@@ -73,6 +123,7 @@ class ZipEADataset(EADataset):
         :param file_name_attr_triples_left: file name of left attribute triples
         :param file_name_attr_triples_right: file name of right attribute triples
         :param file_name_ent_links: file name gold standard containing all entity links
+        :param backend: Whether to use "pandas" or "dask"
         """
         self.zip_path = zip_path
         self.inner_path = inner_path
@@ -81,20 +132,23 @@ class ZipEADataset(EADataset):
         self.file_name_ent_links = file_name_ent_links
         self.file_name_attr_triples_left = file_name_attr_triples_left
         self.file_name_attr_triples_right = file_name_attr_triples_right
+        self.backend = backend
 
         # load data
-        rel_triples_left = self._read_triples(file_name=self.file_name_rel_triples_left)
+        rel_triples_left = self._read_triples(
+            file_name=self.file_name_rel_triples_left, backend=self.backend
+        )
         rel_triples_right = self._read_triples(
-            file_name=self.file_name_rel_triples_right
+            file_name=self.file_name_rel_triples_right, backend=self.backend
         )
         attr_triples_left = self._read_triples(
-            file_name=self.file_name_attr_triples_left
+            file_name=self.file_name_attr_triples_left, backend=self.backend
         )
         attr_triples_right = self._read_triples(
-            file_name=self.file_name_attr_triples_right
+            file_name=self.file_name_attr_triples_right, backend=self.backend
         )
         ent_links = self._read_triples(
-            file_name=self.file_name_ent_links, is_links=True
+            file_name=self.file_name_ent_links, is_links=True, backend=self.backend
         )
         super().__init__(
             rel_triples_left=rel_triples_left,
@@ -104,25 +158,57 @@ class ZipEADataset(EADataset):
             ent_links=ent_links,
         )
 
+    @overload
     def _read_triples(
-        self, file_name: Union[str, pathlib.Path], is_links: bool = False
+        self,
+        file_name: Union[str, pathlib.Path],
+        backend: Literal["pandas"],
+        is_links: bool = False,
     ) -> pd.DataFrame:
+        ...
+
+    @overload
+    def _read_triples(
+        self,
+        file_name: Union[str, pathlib.Path],
+        backend: Literal["dask"],
+        is_links: bool = False,
+    ) -> "dd.DataFrame":
+        ...
+
+    def _read_triples(
+        self,
+        file_name: Union[str, pathlib.Path],
+        backend: BACKEND_LITERAL,
+        is_links: bool = False,
+    ) -> Union[pd.DataFrame, "dd.DataFrame"]:
         columns = (
             list(EA_SIDES) if is_links else (LABEL_HEAD, LABEL_RELATION, LABEL_TAIL)
         )
-        return read_zipfile_csv(
-            path=self.zip_path,
-            inner_path=str(self.inner_path.joinpath(file_name)),
+        read_csv_kwargs = dict(
             header=None,
             names=columns,
             sep="\t",
             encoding="utf8",
             dtype=str,
         )
+        if backend == "pandas":
+            return read_zipfile_csv(
+                path=self.zip_path,
+                inner_path=str(self.inner_path.joinpath(file_name)),
+                **read_csv_kwargs,
+            )
+        else:
+            return read_dask_df_archive_csv(
+                path=self.zip_path,
+                inner_path=str(self.inner_path.joinpath(file_name)),
+                protocol="zip",
+                **read_csv_kwargs,
+            )
 
-    @abstractmethod
     def _param_repr(self) -> str:
-        raise NotImplementedError
+        return f"backend={self.backend}, "
+
 
 class ZipEADatasetWithPreSplitFolds(ZipEADataset):
     """Dataset with pre-split folds created from zip file which is downloaded."""
@@ -136,6 +222,7 @@ class ZipEADatasetWithPreSplitFolds(ZipEADataset):
         file_name_ent_links: str = "ent_links",
         file_name_attr_triples_left: str = "attr_triples_1",
         file_name_attr_triples_right: str = "attr_triples_2",
+        backend: BACKEND_LITERAL = "pandas",
         directory_name_folds: str = "721_5fold",
         directory_names_individual_folds: Sequence[str] = ("1", "2", "3", "4", "5"),
         file_name_test_links: str = "test_links",
@@ -150,6 +237,7 @@ class ZipEADatasetWithPreSplitFolds(ZipEADataset):
         :param file_name_rel_triples_right: file name of right relation triples
         :param file_name_attr_triples_left: file name of left attribute triples
         :param file_name_attr_triples_right: file name of right attribute triples
+        :param backend: Whether to use "pandas" or "dask"
         :param file_name_ent_links: file name gold standard containing all entity links
         :param directory_name_folds: directory name containing folds
         :param directory_names_individual_folds: directory names of individual folds
@@ -165,27 +253,38 @@ class ZipEADatasetWithPreSplitFolds(ZipEADataset):
             file_name_ent_links=file_name_ent_links,
             file_name_attr_triples_left=file_name_attr_triples_left,
             file_name_attr_triples_right=file_name_attr_triples_right,
+            backend=backend,
         )
         self.folds = []
         for fold in directory_names_individual_folds:
             fold_folder = pathlib.Path(directory_name_folds).joinpath(fold)
             train = self._read_triples(
-                fold_folder.joinpath(file_name_train_links), is_links=True
+                fold_folder.joinpath(file_name_train_links),
+                is_links=True,
+                backend=self.backend,
             )
             test = self._read_triples(
-                fold_folder.joinpath(file_name_test_links), is_links=True
+                fold_folder.joinpath(file_name_test_links),
+                is_links=True,
+                backend=self.backend,
             )
             val = self._read_triples(
-                fold_folder.joinpath(file_name_valid_links), is_links=True
+                fold_folder.joinpath(file_name_valid_links),
+                is_links=True,
+                backend=self.backend,
             )
             self.folds.append(TrainTestValSplit(train=train, test=test, val=val))
 
-    def __repr__(self) -> str:
-        len_folds = None if not self.folds else len(self.folds)
-        return f"{self.__class__.__name__}({self._param_repr()}rel_triples_left={len(self.rel_triples_left)}, rel_triples_right={len(self.rel_triples_right)}, attr_triples_left={len(self.attr_triples_left)}, attr_triples_right={len(self.attr_triples_right)}, ent_links={len(self.ent_links)}, folds={len_folds})"
 
 class RDFBasedEADataset(EADataset):
-    def __init__(self, left_file: str, right_file: str, links_file: str, left_format: str, right_format: str):
+    def __init__(
+        self,
+        left_file: str,
+        right_file: str,
+        links_file: str,
+        left_format: str,
+        right_format: str,
+    ):
         logger.info("Loading left graph...")
         left_rel, left_attr = load_from_rdf(left_file, format=left_format)
         logger.info("Loading right graph...")
@@ -202,4 +301,3 @@ class RDFBasedEADataset(EADataset):
     @abstractmethod
     def _load_entity_links(self, ref_path: str) -> pd.DataFrame:
         raise NotImplementedError
-